requirements.txt
Flask
matplotlib
numpy
pandas
python-docx
PyMuPDF
